🧬 Move the factory logic into another new module.
