+++ pymock_api/model/openapi/config.py
@@ -6,7 +6,8 @@ from .. import APIConfig, MockAPI, MockAPIs
 from ..api_config import BaseConfig, _Config
 from ..api_config.apis import APIParameter as PyMockAPIParameter
 from ..enums import ResponseStrategy
-from ._parse import OpenAPIParser, OpenAPIParserFactory, OpenAPIPathParser
+from ._parse import OpenAPIParser, OpenAPIPathParser
+from ._parser_factory import OpenAPIParserFactory
 
 Self = Any
